[Utils] Move external utils to external directory
Some of files in `src/Utils` only consist of external open source codes.
This commit will move them to `src/Utils/external` directory.

ONE-vscode-DCO-1.0-Signed-off-by: Seok NamKoong <seok9311@naver.com>
